Move function outside of file
