+++ pages/cart.tsx
@@ -15,6 +15,7 @@ import {
   getCookies,
   setCookies,
 } from '../utils/cookies';
+import { getSumFromCart } from '../utils/getSumFromCart';
 import { parsePrice } from '../utils/parsePrice';
 
 const cartStyles = css`
@@ -76,19 +77,9 @@ const cartStyles = css`
   }
 `;
 //
-type SpaceshipInCartType = SpaceshipType & {
+export type SpaceshipInCartType = SpaceshipType & {
   quantity: number;
 };
-export function getSumFromCart(spaceships: SpaceshipInCartType[]) {
-  let totalSum = 0;
-  spaceships.map((spaceship) => {
-    if (spaceship.quantity > 0) {
-      return (totalSum += spaceship.quantity * spaceship.price);
-    }
-    return totalSum;
-  });
-  return totalSum;
-}
 
 function increaseQuantity(
   id: number,
